Task 1/requirements.txt
pandas>=2.2.0
openpyxl>=3.1.2
python-docx>=1.1.0
fpdf2>=2.7.8
docx2pdf>=0.1.8
PySide6>=6.10.1
pyinstaller>=6.19.0